requirements.txt
fastapi
uvicorn
torch
torchvision
numpy
opencv-python
pillow
albumentations
tqdm
python-multipart
gdal